requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.23,<1.0
pydantic>=2.6,<3.0
python-dotenv>=1.0,<2.0
httpx>=0.27,<1.0
keybert>=0.8,<0.9
pymupdf>=1.23,<2.0
sentence-transformers>=2.5,<3.0
easyocr>=1.7,<1.8
pillow>=10.2,<11.0
torch>=2.2,<3.0
transformers>=4.38,<5.0
pandas>=2.2,<3.0
numpy>=1.26,<2.0
openpyxl>=3.1,<4.0
